requirements.txt
dotenv
python-dotenv
sqlalchemy
mysql-connector-python
flask_sqlalchemynodemon
pymysql
PyJWT